+++ lib/heatmap_calendar_plus.dart
@@ -4,3 +4,4 @@ export 'src/heatmap.dart';
 export 'src/heatmap_calendar.dart';
 export 'src/data/heatmap_color_mode.dart';
 export 'src/data/heatmap_calendar_type.dart';
+export 'src/data/heatmap_cell_style.dart';